requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
python-dotenv
django-cors-headers